Add External Api Service, Market Companies

+++ ChllengePlusSoft/Models/DesempenhoPostModel.cs
@@ -1,11 +1,11 @@
 ﻿namespace ChllengePlusSoft.Models
 {
     public class DesempenhoPostModel
-    {        
-        public double Receita { get; set; }
-        public double Lucro { get; set; }
-        public double Crescimento { get; set; }
+    {
+        public float Receita { get; set; } 
+        public float Lucro { get; set; }   
+        public float Crescimento { get; set; }
         public long EmpresaId { get; set; }
-    
-}
+    }
+
 }
